river-poultry-backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
Pillow==11.3.0
psycopg2-binary==2.9.10
django-filter==24.3
celery==5.4.0
redis==5.2.1
pywebpush==1.14.0
